interpreter: Handle mixing generated and static rust sources
Rust is more like python than C or C++ in regards to directory struture,
it expects that all sources are in the a single directory tree, and
unlike python it doesn't have something like $PYTHON_PATH or sys.path,
so all of the sources *must* be arranged in a single tree. This becomes
particularly problematic when dealing with a mixture of generated and
static sources.

To deal with this meson will now copy the static sources into the build
directory, using custom_targets. This keeps the source dir clean of
generated cruft, and allows us to build the mixed source targets. By
using custom_targets instead of copying at configure time we avoid the
need to reconfigure when a source changes, instead only one or more
build time copy commands get called, then recompilation. This will
likley be much faster than reconfiguring.

This still doesn't account for generators, which present a different
challenge, they are placed in a per-target private directory, but what
that directory is and were it will be is backend dependant, and this
patch handles all of this in the frontend, not the backend.

Related: #8160