Check custom_target() and run_target() command upfront
If the command has any error, like unsupported argument type, the error
was reported at the end during backend generation, causing the error to
be out of context and complicated to debug for users.

Now check the command from build.CustomTarget() and build.RunTarget()
constructors so the error is raisen immediately.