Fix getting native dependency when doing a native build
When doing a native build, `meson.override_dependency('foo', d)` is
caching the dependency for HOST machine and `dependency('foo', native:
true)` is searching into the cache for BUILD machine. It should make no
difference if `native` kwarg is omitted and when it's set to true when
doing a native build.