dependencies/openmp: Run a compile test
There's a lot more to succesfully using OpenMP than just checking that
the Compiler has a flag, you often need external libraries (libgomp,
libomp, etc). We could attempt to figure out what the Compierl is doing
under the hood for all of that, or we could just do a compile test and
figure out that `-fopenmp` does or does not work.