Fail upon encountering unknown options, add --allow-unknown-options
It is a common error, at least in my experience, that people invoke
"meson configure -D nonexistent=true" and miss the big yellow
uppercase WARNING about "nonexistent" being an unknown option. They
they happily continue building the project in the belief that
"nonexistent" is set to true, when it is, in fact not.

Meson now fails if an unknown option is encountered. The previous
behavior can be restored by using the --allow-unknown-options switch.

Fixes #7288.