mconf: add --fatal-meson-warnings (just as msetup)
It is a common error, at least in my experience, that people invoke
"meson configure -D nonexistent=true" and miss the big yellow
uppercase WARNING about "nonexistent" being an unknown option. They
they happily continue building the project in the belief that
"nonexistent" is set to true, when it is, in fact not.

I am not sure why meason decided that this should only be a
WARNING. But users should at least be able to opt-in for a fail-safe
behavior. This is now possible via the --fatal-meson-warnings
switch for the 'configure' subcommand.

Partially fixes #7288.