interpreter: Correctly store program_overrides per machine
A program for the host machine should only override host machine
binaries, and one for the build machine should only override build
machine binaries.

We don't need a native keyword argument here, because the Targets
already know what machine they're for, and we can use that information.
This is nice from an end user point of view as they don't need to know
whether the binary is host or build, Meson abstracts that away.