Factor out parsing paths from .gitmodules file
This replaces the manual scan for path keys with complete parsing of the file
as dictionaries.  Though overkill at this point, this complete description will
be needed to rewrite the file in the following commit.

Parsing is now done via 'git config', which isolates meson from the details of
the git configuration syntax.  So, a side-effect of this commit should be that
meson is more resilient to unusual files or future possible changes to the
syntax supported by git.

The downside is that this is slower (particularly on Windows where launching a
process is very slow), but this shouldn't be significant compared to everything
else 'meson dist' is doing.