Fix meson dist with relative submodule URLs
Git supports relative submodule URLs, for example "../mylib.git" to refer to a
sibling of the parent project.  This allows mirroring of repositories with
submodules across several hosts.

Configurations like this were broken in 'meson dist', because the submodules
would not be cloned correctly.  This commit checks if any submodules have
relative URLs, and if so, writes a new .gitmodules in the distribution
directory which points to the original submodule path.

In other words, when making a distribution, submodules are cloned in the same
way that the parent project is: from the local git repository.