ninjabackend: Add pch includes as early as possible not to be overridden
When pch are used for a target meson will make the compiler to include
the pre-compiled header. While this is useful, this needs to happen
before any other header has been included, otherwise:

 1) we won't take advantage of pch for anything else previously included
 2) gcc will just fail as it won't even try to look for a pre-compiled
    header in this case [1]

This case can happen quite a easily when a dependency provides an
included header in its cflags.

As per this, split _generate_single_compile() in two phases, one is
responsible of initializing the compiler data, while the other is
defining commands for the context.

In this way, when pch can be used, we can insert the pch inclusion
earlier than any other provided by the target.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=100462