Fix LTO test on Cygwin
This partially reverts commit add502c6483bde9dc6a0ba80b3c79163304465a4.

In 'linkshared' test, annotate cppfunc() as imported, so an indirection
through an import stub is generated, avoiding a relocation size error
when building using gcc for Cygwin with LTO on.

Align with the example of how to write this portably in [1].

The 'c' language part of that test already gets this right.

[1] http://gcc.gnu.org/wiki/Visibility