Set MESON_CI_JOBNAME for all CI jobs
Set MESON_CI_JOBNAME for all CI jobs which run project tests.

(Note that ${{ github.job }} is the literal job.id used in the yaml, not
any name given to the job with job.id.name, and so is the same for all
matrix entries, and thus not suitable for our purposes there).