Skip LLVM test if required modules aren't found
If the required LLVM modules can't be found, skip the LLVM framework
test, rather than succesfully doing nothing.

(This optionality is a leftover from before #7379)

(At the moment, OpenSuse provides dynamic-only LLVM.  The cmake method
still finds LLVM, when a static LLVM is requested, but fails to find any
modules.  This might be a bug in the cmake method of the LLVM
dependency.)