Fix skip condition in sdl2 test.json
The build_machine.system() value we should be matching against is
'darwin', not 'macos', so this was expected to skip everywhere.

Unfortunately, fixing this reveals that this test skips in our macos CI
runs also, (I think) because brew doesn't (normally) provide frameworks,
just pkgconfig files.

So, also skip this test there, which means it doesn't run anywhere in
CI currently.