coverage: prefer gcovr over lcov for html
The documentation in Feature-autodetection.md[1] implies that `gcovr`
is favored over `lcov`, but that is not the current implementation.
It was reported in #6747 that `lcov` output is incorrect when test
cases are present in a subdirectory, but `gcovr` yields the correct
results.  Swap the preference when both are found so that the better
tool is used.

1. https://github.com/mesonbuild/meson/blame/b20cfec4bc9f540cb19e86c97014776be0197eff/docs/markdown/Feature-autodetection.md#L27

Partially fixes mesonbuild/meson#6747.
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>