compilers: allow link tests to use objects from a different compiler
In some cases, link tests would like to use objects provided by a compiler
for a different language, for example linking a C object file with a C++
compiler.  This kind of scenario is what link_language is for, but it is
impossible to test that it works with a linker test.

This patch implements the required support in the Compiler class.  The
source code compiler is passed to the Compiler.links method as an
argument.