Fix when gtest dep is found but does not pass version check
GTestDependencySystem (and other similar dep classes) sets
self.is_found=True, but the version check could still fail. In the case
the dependency is not required `ExternalDependency._check_version()`
won't raise an exception and thus the dependency is accepted.

_check_version() sets self.is_found() in the case self.version is not
empty, we should do it too when self.version is empty.

Fixes: #9036.